fix a warning

+++ cpp/include/patterns/update_frontier_v_push_if_out_nbr.cuh
@@ -598,8 +598,8 @@ void update_frontier_v_push_if_out_nbr(
                                                         (1 + tuple_size));
     for (size_t i = 0; i < tx_counts.size(); ++i) {
       auto comm_dst_rank = graph_view.is_hypergraph_partitioned()
-                             ? col_comm_rank * row_comm_size + i
-                             : row_comm_rank * col_comm_size + i;
+                             ? col_comm_rank * row_comm_size + static_cast<int>(i)
+                             : row_comm_rank * col_comm_size + static_cast<int>(i);
       if (comm_dst_rank == comm_rank) {
         assert(i == tx_self_i);
         // FIXME: better define request_null (similar to MPI_REQUEST_NULL) under raft::comms
@@ -623,8 +623,8 @@ void update_frontier_v_push_if_out_nbr(
     }
     for (size_t i = 0; i < rx_counts.size(); ++i) {
       auto comm_src_rank = graph_view.is_hypergraph_partitioned()
-                             ? col_comm_rank * row_comm_size + i
-                             : row_comm_rank + i * row_comm_size;
+                             ? col_comm_rank * row_comm_size + static_cast<int>(i)
+                             : row_comm_rank + static_cast<int>(i) * row_comm_size;
       if (comm_src_rank == comm_rank) {
         assert(self_tx_i != std::numeric_limits<size_t>::max());
         assert(rx_counts[i] == tx_counts[tx_self_i]);
